Conflicts:
	www/index.html

+++ src/tools/haxelib/compatible/Convert.hx
@@ -13,30 +13,30 @@ class Convert {
 
   public static
   function main() {
-    
+
     var
-      users = getUsers(),
+      //users = getUsers(),
       client = new ClientRestful(),
       tmpDir = "/tmp/unzip/",
-      fs = Os.files("/home/blackdog/OLDHAXLIB"), // the haxelib zip directory
+      fs = Os.files("/home/daniel/oldhaxe"), // the haxelib zip directory
       nnull = 0,
       i = 0;
 
     for (f in fs) {
-      Os.rmdir(tmpDir);
+      if (Os.exists(tmpDir)) Os.rmdir(tmpDir);
       Os.mkdir(tmpDir);
-        
+
       var
         zf = neko.io.File.read(f,true),
         zip = neko.zip.Reader.readZip(zf);
-        
+
       Os.unzip(zip,tmpDir);
 
       var
         xmlFile = find(tmpDir,["-name","haxelib.xml"])[0],
         xml = Os.fileIn(xmlFile),
         newPackage;
-        
+
       if (xml == null) {
         trace("null in "+f);
         nnull++;
@@ -45,12 +45,13 @@ class Convert {
       var
         data = Datas.readData(xml),
         dev = data.developers.first(),
-        user = users.get(dev),
+        //user = users.get(dev),
+        user = { pw: "12345", fn: "Old Haxelib" },
         o = new Options(),
         email = dev+"@haxe.org";
 
-      o.addSwitch("-R","lib.ipowerhouse.com");
-      
+      o.addSwitch("-R","localhost:8200");
+
       client.register(o,email,user.pw,user.fn,
                       function(rurl:String,s:Status) {
                         trace("registed "+dev+"@haxe.org");
@@ -58,33 +59,33 @@ class Convert {
                       });
 
       Reflect.setField(data,"email",email);
-      Os.fileOut(tmpDir+"/haxelib.json",toHpx(data));
+      Os.fileOut(tmpDir+"/haxelib.json",toHxp(data));
       newPackage = packit(tmpDir+"/haxelib.json");
-    
+
       zf.close();
-    
+
       client.submit(o,user.pw,newPackage,function(rurl:String,s:Status) {
           trace("submitted to "+rurl);
           trace("status "+s);
           return true;
         });
-      
+
       if (i++ == 20) break;     // don't do them all right now
     }
-      
+
     trace("nnull = "+nnull);
   }
-
+/*
   public static function
   getUsers():Hash<OldUser> {
     var
       sql = "select name,pass,fullname from user",
       users = new Hash<OldUser>(),
       rows;
-    
+
     cnx = neko.db.Sqlite.open("haxelib.db");
     rows = cnx.execute(sql);
-    
+
     for (r in rows) {
       users.set(r.name,{pw:r.pass,fn:r.fullname});
     }
@@ -92,21 +93,44 @@ class Convert {
     cnx.close();
     return users;
   }
-  
+*/
   public static function
-  toHpx(x:XmlInfos) {
+  toHxp(x:XmlInfos) {
+    var data: Dynamic = {}, d: Dynamic, t: Dynamic;
+    for (field in Reflect.fields(x)) {
+      t = Reflect.field(x, field);
+      if (Std.is(t, List) && (t.length == 0 || Std.is(t.first(), String))) {
+        d = t.join(" ");
+      } else if (Std.is(t, String) && cast(t, String).indexOf("\n") > -1) {
+        d = "";
+        var a: Array<String> = StringTools.trim(t).split("\n");
+        var f = true;
+        for (s in a) {
+          if (f) {
+            f = false;
+            d += s;
+          } else {
+            d += "\n                " + StringTools.ltrim(s);
+          }
+        }
+      } else {
+        d = Std.string(t);
+      }
+      if (d.length == 0) d = "null";
+      Reflect.setField(data, field, d);
+    }
     var tmpl = '
-name:           ::project::
+project:        ::project::
 website:        ::website::
 version:        ::version::
 comments:       ::versionComments::
 description:    ::desc::
 author-email:   ::email::
-tags:           ::foreach tags::::name::::end::
+tags:           ::tags::
 license:        ::license::
-author:         ::foreach developers::::name::::end::
+author:         ::developers::
 ';
-    return Os.template(tmpl,x);
+    return Os.template(tmpl,data);
     //build-depends: ::foreach dependencies::::project:: >= ::version:: ::end::
 
   }
@@ -121,7 +145,7 @@ author:         ::foreach developers::::name::::end::
 		var o = new neko.io.Process("find",prms).stdout;
 		return textToArray(o.readAll().toString());
 	}
-    
+
 	public static
 	function textToArray(text:String,delimiter='\n'):Array<String> {
 		var ar = text.split(delimiter),
@@ -133,12 +157,12 @@ author:         ::foreach developers::::name::::end::
 	}
 
     public static function
-    packit(hpxFile:String) {
+    packit(hxpFile:String) {
       var
-        hpx = Parser.process(hpxFile),
-        conf = Parser.getConfig(hpx);
-      
+        hxp = Parser.process(hxpFile),
+        conf = Parser.getConfig(hxp);
+
       return Package.createFrom(conf);
-    }  
+    }
 
 }
